requirements.txt
# Core dependencies
absl-py==2.1.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
biopython==1.79
biotite==0.40.0
blosum==2.0.2
cachetools==5.5.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
contextlib2==21.6.0
cycler==0.11.0
deepspeed==0.5.9
dm-tree==0.1.6
docker-pycreds==0.4.0
einops==0.6.1
exceptiongroup==1.1.1
fairscale==0.4.13
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
google-auth-oauthlib==0.4.6
greenlet==3.1.1
grpcio==1.62.3
hjson==3.1.0
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
ml-collections==0.1.0
msgpack==1.0.5
multidict==6.0.5
networkx==2.6.3
ninja==1.11.1.3
numba==0.56.4
numpy==1.21.2
oauthlib==3.2.2
openmm==7.5.1
optuna==4.0.0
pandas==1.0.5
pathtools==0.1.2
Pillow==9.3.0
promise==2.3
protobuf==3.20.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyDeprecate==0.3.1
pymol==2.4.0
pyparsing==3.1.4
PyQt5==5.15.7
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2023.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.1
sentry-sdk==2.19.2
setproctitle==1.3.3
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.41
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.11.4
torchvision==0.13.1
tqdm==4.62.2
triton==1.0.0
typing_extensions==4.7.1
urllib3==1.26.13
wandb==0.12.21
Werkzeug==2.2.3
yarl==1.9.4

# Local development packages (must be installed manually or cloned from repository)
# fair-esm: https://github.com/facebookresearch/esm
# DLLogger: https://github.com/pytorch/ELF/tree/main/dllogger
# openfold: https://github.com/aqlaboratory/openfold
# ESMFold: follow installation instructions from https://github.com/facebookresearch/esm
